+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Add filter details for the "view all" log screen (UIDATIMP-287)
 * Update query for sorting by numeric field (UIDATIMP-310)
 * Update dependency on source-record-manager (UIDATIMP-318)
+* Remove "React to" from action profiles(UIDATIMP-316)
 
 ## [1.5.1](https://github.com/folio-org/ui-data-import/tree/v1.5.1) (2019-09-25)
 * Create Section component (UIDATIMP-278)